move reporter internals in both headers and source
