+++ src/CMakeLists.txt
@@ -3,7 +3,8 @@ include_directories(${PROJECT_SOURCE_DIR}/src)
 
 # Define the source files
 set(SOURCE_FILES "benchmark.cc" "colorprint.cc" "commandlineflags.cc" "log.cc"
-                 "sleep.cc" "string_util.cc" "sysinfo.cc" "walltime.cc")
+                 "reporter.cc" "sleep.cc" "string_util.cc" "sysinfo.cc"
+                 "walltime.cc")
 # Determine the correct regular expression engine to use
 if(HAVE_STD_REGEX)
   set(RE_FILES "re_std.cc")
